+++ sr_gui_fingertip_visualization/src/sr_gui_fingertip_visualization/tab_layouts_graph.py
@@ -21,7 +21,7 @@ import rospy
 import rospkg
 from sr_robot_msgs.msg import ShadowPST, BiotacAll
 from sr_gui_fingertip_visualization.generic_plots import GenericDataPlot
-from sr_gui_fingertip_visualization.tab_layouts_generic import GenericGraphTab
+from sr_gui_fingertip_visualization.tab_layouts_generic import GenericGraphTab, GenericOptionBar, GenericOptionBar
 
 
 from python_qt_binding.QtCore import QTimer
@@ -111,7 +111,7 @@ class GraphTab(QWidget):
                 self.tactile_widgets[side] = BiotacGraphTab(side, self)
             self.stacked_layout.addWidget(self.tactile_widgets[side])
 
-        self._option_bar = OptionBar(list(self._tactile_topics.keys()), childs=self.stacked_layout)
+        self._option_bar = GraphOptionBar(list(self._tactile_topics.keys()), childs=self.stacked_layout)
         
         finger_layout.addWidget(self._option_bar)        
         finger_layout.addLayout(self.stacked_layout)
@@ -121,73 +121,19 @@ class GraphTab(QWidget):
         return self.tactile_widgets
 
 
-class OptionBar(QGroupBox):
+class GraphOptionBar(GenericOptionBar):
     def __init__(self, hand_ids, childs):
-        super().__init__()
-        self._childs = childs
-        self._fingers = ["ff", 'mf', 'rf', 'lf', 'th']
+        super().__init__(hand_ids, childs)
+        self.init_layout()
+        self.create_connections()
 
-        self.setTitle("Options")
-        self.setSizePolicy(1, 2)
+    def init_layout(self):
+        super().init_layout()
+        self.options_layout.addLayout(self.hand_id_selection_layout)
+        self.options_layout.addStretch(1)
+        self.options_layout.addWidget(self.finger_selection_label)
+        self.options_layout.addWidget(self.finger_selection_show_selected_button)
+        self.options_layout.addWidget(self.finger_selection_show_all_button)
 
-        options_layout = QHBoxLayout()
-
-        hand_id_selection_layout = QFormLayout()
-        self.hand_id_selection = QComboBox()
-        self.hand_id_selection.addItems(hand_ids)
-        hand_id_selection_layout.addRow(QLabel("Hand ID:"), self.hand_id_selection)
-
-        finger_selection_label = QLabel("Finger selection:")
-        self.finger_selection_show_selected_button = QPushButton("Show selected")
-        self.finger_selection_show_selected_button.setSizePolicy(2, 2)
-        self.finger_selection_show_all_button = QPushButton("Show all")
-        self.finger_selection_show_all_button.setSizePolicy(2, 2)
-
-        options_layout.addLayout(hand_id_selection_layout)
-        options_layout.addStretch(1)
-        options_layout.addWidget(finger_selection_label)
-        options_layout.addWidget(self.finger_selection_show_selected_button)
-        options_layout.addWidget(self.finger_selection_show_all_button)
-
-        self.setLayout(options_layout)
+        self.setLayout(self.options_layout)
         self._current_widget = self._childs.currentWidget()        
-        self._childs.setCurrentIndex(self.hand_id_selection.currentIndex())
-        self._create_connections()
-
-    def _create_connections(self):
-        self.hand_id_selection.currentIndexChanged.connect(self._combobox_action_hand_id_selection)
-        self.finger_selection_show_selected_button.clicked.connect(self._button_action_show_selected_fingers)
-        self.finger_selection_show_all_button.clicked.connect(self._button_action_show_all)
-
-    def _combobox_action_hand_id_selection(self):
-        self._current_widget = self._childs.currentWidget()        
-        self._start_selected_widget(self._current_widget)
-        self._childs.setCurrentIndex(self.hand_id_selection.currentIndex())
-
-    def _button_action_show_selected_fingers(self):
-        fingertip_widgets = self._childs.currentWidget().get_finger_widgets()
-        self._selected_fingers = [finger for finger in self._fingers if fingertip_widgets[finger].isChecked()]
-        for finger in self._fingers:
-            if finger in self._selected_fingers:
-                fingertip_widgets[finger].start_timer_and_subscriber()
-                fingertip_widgets[finger].show()
-            else:
-                fingertip_widgets[finger].stop_timer_and_subscriber()
-                fingertip_widgets[finger].hide()
-
-    def _button_action_show_all(self):
-        fingertip_widgets = self._childs.currentWidget().get_finger_widgets()
-        self._selected_fingers = [finger for finger in self._fingers if fingertip_widgets[finger].isChecked()]
-        for finger in self._fingers:
-            fingertip_widgets[finger].setChecked(False)
-            fingertip_widgets[finger].stop_timer_and_subscriber()
-            fingertip_widgets[finger].show()
-
-    def _start_selected_widget(self, selected_widget):
-        for widget_index in range(self._childs.count()):
-            finger_widgets_from_tab = self._childs.currentWidget().get_finger_widgets()
-            for finger, widget in finger_widgets_from_tab.items():
-                if self._childs.currentWidget() == selected_widget:
-                    widget.start_timer_and_subscriber()
-                else:
-                    widget.start_timer_and_subscriber()
